# Performance improvements: moving from data organized in an Array of Structures (AoS) to an organization of Stucture of Arrays (SoA)