@filipecosta90 2021-02-22
- add the hot code for `td_add`. 
- prove that that portion of code is not vectored on master
- prove that with the array of primary data-types is vectorized
- get numbers for `td_quantile` on master 
- get numbers for `td_cdf` on master 
- get numbers for `merge` on master