@filipecosta90 2021-08-07
@ashtul merging given I need the most performant version for a multi sketch benchmark